+++ Jellyfin.Api/Controllers/BrandingController.cs
@@ -3,6 +3,7 @@ using MediaBrowser.Controller.Configuration;
 using MediaBrowser.Model.Branding;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Jellyfin.Api.Controllers;
 
@@ -27,6 +28,7 @@ public class BrandingController : BaseJellyfinApiController
     /// </summary>
     /// <response code="200">Branding configuration returned.</response>
     /// <returns>An <see cref="OkResult"/> containing the branding configuration.</returns>
+    [AllowAnonymous]
     [HttpGet("Configuration")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     public ActionResult<BrandingOptionsDto> GetBrandingOptions()
@@ -52,6 +54,7 @@ public class BrandingController : BaseJellyfinApiController
     /// An <see cref="OkResult"/> containing the branding css if exist,
     /// or a <see cref="NoContentResult"/> if the css is not configured.
     /// </returns>
+    [AllowAnonymous]
     [HttpGet("Css")]
     [HttpGet("Css.css", Name = "GetBrandingCss_2")]
     [Produces("text/css")]