+++ Jellyfin.Api/Controllers/DashboardController.cs
@@ -70,6 +70,7 @@ public class DashboardController : BaseJellyfinApiController
     /// <response code="200">ConfigurationPage returned.</response>
     /// <response code="404">Plugin configuration page not found.</response>
     /// <returns>The configuration page.</returns>
+    [AllowAnonymous]
     [HttpGet("web/ConfigurationPage")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]