+++ Jellyfin.Api/Controllers/SystemController.cs
@@ -89,6 +89,7 @@ public class SystemController : BaseJellyfinApiController
     /// </summary>
     /// <response code="200">Information retrieved.</response>
     /// <returns>A <see cref="PublicSystemInfo"/> with public info about the system.</returns>
+    [AllowAnonymous]
     [HttpGet("Info/Public")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     public ActionResult<PublicSystemInfo> GetPublicSystemInfo()
@@ -99,6 +100,7 @@ public class SystemController : BaseJellyfinApiController
     /// </summary>
     /// <response code="200">Information retrieved.</response>
     /// <returns>The server name.</returns>
+    [AllowAnonymous]
     [HttpGet("Ping", Name = "GetPingSystem")]
     [HttpPost("Ping", Name = "PostPingSystem")]
     [ProducesResponseType(StatusCodes.Status200OK)]